Feat/#102-알림 토픽 발행 및 DTO 타입 변경

+++ ET-core/src/main/java/com/team1/etcore/trade/dto/SettlementDTO.java
@@ -17,6 +17,6 @@ public class SettlementDTO {
     private String stockCode;
     private Position position;
     private BigDecimal orderPrice;
-    private int orderAmount;
+    private BigDecimal orderAmount;
     private Position orderPosition;
 }
